# Improve performance of LookupCommandInfo.