Converting command names to upper case is a non-trival part of our
profile. Assuming that 99% of programmers don't use mixed case,
we can optimize that by explicitly checking the lower/upper case
commands and short-circuiting before upper casing.

This makes mixed cases marginally slower, but makes the common case much
faster. A good trade-off I think.

```
BEFORE:

BenchmarkLookupCommandInfoCorrectCase    2000000           716 ns/op
BenchmarkLookupCommandInfoMixedCase      3000000           532 ns/op

AFTER:

BenchmarkLookupCommandInfoCorrectCase   10000000           131 ns/op
BenchmarkLookupCommandInfoMixedCase     2000000           593 ns/op
```