@garyburd 2015-06-24
Take a look at https://github.com/garyburd/redigo/compare/lookup-perf.  I think it will match the performance of this PR, but be more maintainable when all commands are added to support Redis cluster.