@clutchski 2015-06-24
:+1: as long as the benchmark looks good, seems like a fine solution to me.