@clutchski 2015-06-24
also .. i have some crap in the benchmark file (e.g. unrelated comments) and a wrong err msg which i just realized should be cleaned up.