@garyburd 2015-06-24
Perf is identical for matching case.

    $ git checkout clutchski-performance
    Switched to branch 'clutchski-performance'
    $ go test -test.bench=.
    PASS
    BenchmarkLookupCommandInfoCorrectCase   20000000           104 ns/op
    BenchmarkLookupCommandInfoMixedCase  3000000           519 ns/op
    ok      github.com/garyburd/redigo/internal 4.276s
    $ git checkout lookup-perf
    Switched to branch 'lookup-perf'
    $ go test -test.bench=.
    PASS
    BenchmarkLookupCommandInfoCorrectCase   20000000           102 ns/op
    BenchmarkLookupCommandInfoMixedCase  3000000           535 ns/op
    ok      github.com/garyburd/redigo/internal 4.293s

I'll cleanup the comments and submit it.