@garyburd 2015-06-24
Thank you for pointing out the issue.  See 27a6a7d7c76db7d75feb28f82e91cb34160606c3.